+++ extension/src/popup/components/hardwareConnect/LedgerSign/styles.scss
@@ -1,4 +1,5 @@
-.LedgerConnect {
+// ALEC TODO - change to LedgerSign?
+.LedgerSign {
   z-index: var(--z-index--ledger);
   position: fixed;
   height: var(--popup--height);
