+++ extension/src/popup/constants/routes.ts
@@ -6,6 +6,7 @@ export enum ROUTES {
   importAccount = "/account/import",
   connectWallet = "/account/connect",
   connectWalletPlugin = "/account/connect/plugin",
+  connectLedger = "/account/connect/ledger",
   accountHistory = "/account-history",
   sendPayment = "/account/sendPayment",
   sendPaymentTo = "/account/sendPayment/to",
