+++ extension/src/popup/ducks/transactionSubmission.ts
@@ -226,6 +226,8 @@ interface TransactionData {
 
 interface HardwareWalletData {
   status: HwOverlayStatus;
+  // ALEC TODO - just have 1 status obv
+  signingStatus: HwOverlayStatus;
   transactionXDR: string;
 }
 
@@ -260,6 +262,7 @@ const initialState: InitialState = {
   },
   hardwareWalletData: {
     status: HwOverlayStatus.IDLE,
+    signingStatus: HwOverlayStatus.IDLE,
     transactionXDR: "",
   },
   accountBalances: {
@@ -320,6 +323,7 @@ const transactionSubmissionSlice = createSlice({
       state.hardwareWalletData.status = HwOverlayStatus.IDLE;
       state.hardwareWalletData.transactionXDR = "";
     },
+    
   },
   extraReducers: (builder) => {
     builder.addCase(submitFreighterTransaction.pending, (state) => {
@@ -379,6 +383,16 @@ const transactionSubmissionSlice = createSlice({
       state.transactionData.path = path;
       state.transactionData.destinationAmount =
         action.payload.destination_amount;
+    });
+    builder.addCase(signWithLedger.pending, (state) => {
+      state.hardwareWalletData.signingStatus = HwOverlayStatus.IN_PROGRESS;
+    });
+    // ALEC TODO - fix these two below, obv
+    builder.addCase(signWithLedger.rejected, (state) => {
+      state.hardwareWalletData.signingStatus = HwOverlayStatus.IDLE;
+    });
+    builder.addCase(signWithLedger.fulfilled, (state) => {
+      state.hardwareWalletData.signingStatus = HwOverlayStatus.IDLE;
     });
   },
 });
